# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: StayKey API
  version: 0.1.0
  description: |
    Contract for the StayKey backend. Owner endpoints live under /v1, public
    endpoints used by booking pages and the widget live under /v1/public.
    Money is always an integer amount in minor units (cents) with an ISO currency.

    Owner routes take a bearer access token and an X-Account-Id header naming the
    account to act on. A token for someone who is not a member of that account gets
    404, the same as an account that does not exist.
servers:
  - url: https://api.staykey.direct
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/properties:
    get:
      tags: [properties]
      operationId: listProperties
      summary: List the properties the caller can see
      description: Caretakers limited to some properties only see those.
      security:
        - bearerAuth: []
          accountHeader: []
      responses:
        "200":
          description: Properties, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [properties]
      operationId: createProperty
      summary: Create a property and its booking page address
      description: Owners only, since a property counts towards the plan.
      security:
        - bearerAuth: []
          accountHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewProperty"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The booking page address is already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from /v1/auth/verify or /v1/auth/refresh. Lasts 15 minutes.
    accountHeader:
      type: apiKey
      in: header
      name: X-Account-Id
      description: The account an owner route acts on. The caller must be an active member.
  responses:
    Unauthorized:
      description: The access token is missing, invalid, expired or revoked
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The caller's role does not allow this
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    BadRequest:
      description: The request was invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Slug:
      type: string
      description: Lowercase subdomain label used for the booking page, for example "kingfisher" in kingfisher.staykey.direct.
      pattern: "^[a-z0-9](?:[a-z0-9-]{1,38}[a-z0-9])$"
      example: kingfisher
    BookingType:
      type: string
      description: Whether guests book the whole place or individual rooms.
      enum: [entire, rooms]
    Currency:
      type: string
      enum: [USD, LKR, EUR, GBP]
    NewProperty:
      type: object
      required: [name, bookingType, currency, baseRate]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
          example: Kingfisher Villa
        slug:
          $ref: "#/components/schemas/Slug"
        bookingType:
          $ref: "#/components/schemas/BookingType"
        location:
          type: string
          maxLength: 160
          example: Unawatuna, Galle
        currency:
          $ref: "#/components/schemas/Currency"
        baseRate:
          type: integer
          format: int64
          minimum: 1
          description: Nightly base rate in minor units.
          example: 18000
    Property:
      type: object
      required: [id, slug, name, bookingType, currency, baseRate, bookingPageUrl, createdAt]
      properties:
        id:
          type: string
          format: uuid
        slug:
          $ref: "#/components/schemas/Slug"
        name:
          type: string
        bookingType:
          $ref: "#/components/schemas/BookingType"
        location:
          type: string
        currency:
          $ref: "#/components/schemas/Currency"
        baseRate:
          type: integer
          format: int64
        bookingPageUrl:
          type: string
          format: uri
          example: https://kingfisher.staykey.direct
        createdAt:
          type: string
          format: date-time
    PropertyList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Property"
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: slug_taken
        message:
          type: string
          example: That booking page address is already taken.
        field:
          type: string
          description: The request field the error is about, when there is one.
          example: slug
        retryAfter:
          type: integer
          description: Seconds to wait before retrying, on 429 responses.